+++ packages/components/modules/profiles/web/ProfileComponent/ReportButtonWithDialog/constants.ts
@@ -1,0 +1,7 @@
+export const STEPS = {
+  report: 'report',
+  subTypes: 'subTypes',
+  text: 'text',
+  summary: 'summary',
+  confirmation: 'confirmation',
+}